requirements.txt
streamlit
supabase
google-generativeai
python-dotenv
feedparser
pandas
requests
pytest
pytest-mock
plotly
openpyxl
